Move annotator.plugin.editor -> annotator.ui.editor.standalone
